% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.DIA}
\alias{data.DIA}
\title{Example DIA proteomics data}
\format{
## `data.DIA`
A data frame with 2000 rows and 100 columns:
\describe{
  \item{study}{study name}
  \item{year}{2025}
  ...
}
}
\source{
NCI-CPTAC
}
\usage{
data.DIA
}
\description{
A subset of data from a DIA proteomics experiment
}
\keyword{datasets}
