% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_imputation_functions.R
\name{gamma_est}
\alias{gamma_est}
\title{Estimate abundance dependent missing parameter}
\usage{
gamma_est(data)
}
\arguments{
\item{data}{dataset in the form of a matrix or dataframe with missing values or NA's.}
}
\value{
intercept and slope of the log linear trend between missing rate and protein mean abundance.
}
\description{
The function estimate log linear dependency between missing rate and protein abundance.
}
\details{
Input of this function is the data matrix with rows as proteins and columns as samples. Out put is the parameter of abundance missing dependence.
}
\examples{
\dontrun{
data<-data.DIA[1:100,1:50]
gm_temp = gamma_est(data)
}
}
