% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_imputation_functions.R
\name{impute.mice}
\alias{impute.mice}
\title{Imputation of Missing Protein Abundances using MICE}
\usage{
impute.mice(data, m = 1, method = "pmm", maxit = 20, seed, ...)
}
\arguments{
\item{data}{dataset in the form of a matrix or data frame with NAs as missings.}

\item{m}{Number of multiple imputations. The default is m=1.}

\item{method}{Specifying the imputation method to be used for each column in data. The default is 'pmm'.}

\item{maxit}{A scalar giving the number of iterations. The default is 20.}

\item{seed}{random seed used by the random number generator set.seed().}

\item{...}{other parameters in mice() function}
}
\value{
the imputed version of the dataset
}
\description{
The function impute.mice imputes a dataset with missing values or NA's using mice.
}
\examples{
\dontrun{
data<-data.DIA[1:100,1:50]
impute.mice(data=as.matrix(data))
}
}
