% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroFM.R
\name{impute.SpectroFM}
\alias{impute.SpectroFM}
\title{Imputation of Missing Protein Abundances using Matrix Factorization}
\usage{
impute.SpectroFM(input_table, iter, verbose = FALSE)
}
\arguments{
\item{input_table}{A data frame with missing values or Nas}

\item{iter}{Number of iterations}

\item{verbose}{verbose in mcmc procedure}
}
\value{
Imputed data frame
}
\description{
The function impute.SpectroFM imputes a dataset with missing values or NA's using factorization
}
\examples{
\dontrun{
data<-data.DIA[1:100,1:50]
impute.SpectroFM(input_table=as.data.frame(data), iter = 40)
}
}
